Split rule tests by area
